nah keep it in same file
