+++ ast/ast.go
@@ -30,7 +30,27 @@ type Identifier struct {
 	Token token.Token
 	Value string
 }
-
 func (i *Identifier) expressionNode() {}
-
 func (i *Identifier) TokenLiteral() string { return i.Token.Literal }
+
+type VarStatement struct {
+	Token token.Token
+	Name  *Identifier
+	Value Expression
+}
+func (ls *VarStatement) statementNode() {}
+func (ls *VarStatement) TokenLiteral() string { return ls.Token.Literal }
+
+type ReturnStatement struct {
+	Token       token.Token
+	ReturnValue Expression
+}
+func (rs *ReturnStatement) statementNode()       {}
+func (rs *ReturnStatement) TokenLiteral() string { return rs.Token.Literal }
+
+type ExpressionStatement struct {
+	Token token.Token
+	Expression Expression
+}
+func (es *ExpressionStatement) statementNode() {}
+func (es *ExpressionStatement) TokenLiteral() string { return es.Token.Literal }
